Added registration and login backend functionality
Several packages were pulled in to accomplish the functionality, including:
-flask_bcrypt (password hashing)
-flask_jwt_extended (jwt functionality)
Validation was added to the username and login_email fields in the user model for data integrity. The server Dockerfile was changed to check for any changes to the db_models, generate new migration scripts, and update the database on container startup.